requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
markdown>=3.5
